ASL-4060 - Switch to optimised query for internal-deadlines
Adjustment made to the test setup to ensure the activity is added to the task in date order as this logic is now in the query.
