+++ lib/reports/internal-deadlines/index.js
@@ -43,13 +43,19 @@ module.exports = ({ db, query: params }) => {
     const q = db.flow('cases')
       .select([
         'cases.*',
-        db.flow.raw('JSON_AGG(activity_log.*) as activity')
+        db.flow.raw('JSON_AGG(activity_log.* ORDER BY activity_log.created_at) as activity')
       ])
       .leftJoin('activity_log', 'cases.id', 'activity_log.case_id')
       .whereRaw(`cases.data->>'model' = 'project'`)
       .whereRaw(`cases.data->>'action' = 'grant'`)
       .where('cases.updated_at', '>', earliest)
       .where('cases.created_at', '<=', latest)
+      .where(orBuilder => {
+        let statuses = [...targetClearingStatuses, ...closedStatuses];
+        return orBuilder.where(withinDateBuilder => withinDateBuilder.where('activity_log.updated_at', '>', earliest).where('activity_log.created_at', '<=', latest))
+          .orWhere(statusBuilder => statusBuilder.whereRaw(`activity_log.event->>'status' in (${statuses.map(_ => '?').join(',')})`, statuses));
+      })
+      .where(builder => builder.where('activity_log.event_name', '=', 'update').orWhere('activity_log.event_name', 'like', 'status:%'))
       .groupBy('cases.id');
 
     return q;
@@ -64,8 +70,6 @@ module.exports = ({ db, query: params }) => {
     let extended;
 
     record.activity
-      .filter(a => a.event_name.match(/^status:/) || a.event_name === 'update')
-      .sort((a, b) => a.created_at < b.created_at ? -1 : 1) // ascending time order
       .forEach(activity => {
         if (!target) {
           const internalDeadline = get(activity, 'event.data.internalDeadline');
